refactor: Extend Probot class
