+++ src/index.ts
@@ -1,110 +1,8 @@
-import { Probot } from 'probot';
-import fs from 'fs';
+import { ProbotOnKube } from './probot';
 import { V1Secret } from '@kubernetes/client-node';
-const { createAppAuth } = require('@octokit/auth-app');
-const k8s = require('@kubernetes/client-node');
-require('dotenv').config();
+import { InstallationAccessTokenAuthentication } from '@octokit/auth-app';
 
-if (process.env.NODE_ENV !== 'production') {
-  dotenv.config();
-}
-
-const EXPIRATION_THRESHOLD = 5 * 60000;
-
-// K8s client for using k8s apis.
-const kc = new k8s.KubeConfig();
-kc.loadFromDefault();
-const k8sApi = kc.makeApiClient(k8s.CoreV1Api);
-const k8sCustomResourceApi = kc.makeApiClient(k8s.CustomObjectsApi);
-const k8sContext = kc.getCurrentContext();
-const k8sNamespace = (() => {
-  if (k8sContext === 'inClusterContext') {
-    return fs.readFileSync(
-      '/var/run/secrets/kubernetes.io/serviceaccount/namespace',
-      'utf8'
-    );
-  } else {
-    return k8sContext.split('/')[0];
-  }
-})();
-console.info(`Operating in the namespace: ${k8sNamespace}`);
-
-/**
- * Function that patches K8s secret with updated token and expiry time
- * @param    {Number} installId    Installation ID
- * @return   None
- */
-async function checkExpiryUpdateToken(installId: number) {
-  const appSecret = await k8sApi.readNamespacedSecret(
-    'peribolos-' + installId,
-    k8sNamespace
-  );
-  if (!appSecret) {
-    console.error('Secret not found');
-    return;
-  }
-
-  console.info(
-    `Working on the Installation: ${appSecret.body?.metadata?.name}`
-  );
-  const current_date = new Date();
-  const expiry_date = new Date(
-    appSecret.body?.metadata?.annotations?.expiresAt || 0
-  );
-
-  // check if token not expired
-  if (expiry_date.getTime() > current_date.getTime() + EXPIRATION_THRESHOLD) {
-    console.info(
-      `Token is still fresh for Installation: ${appSecret?.body?.metadata?.name}`
-    );
-    return;
-  }
-
-  console.info(
-    `Token is expired for Installation: ${appSecret?.body?.metadata?.name}`
-  );
-  const auth = await createAppAuth({
-    appId: process.env.APP_ID as string,
-    privateKey: process.env.PRIVATE_KEY as string,
-  });
-  const resToken = await auth({
-    type: 'installation',
-    installationId: installId,
-  });
-
-  const secret: V1Secret = {
-    metadata: {
-      annotations: {
-        expiresAt: resToken.expiresAt,
-      },
-    },
-    stringData: {
-      token: resToken.token,
-    },
-  };
-  const headers = { 'content-type': 'application/merge-patch+json' };
-  k8sApi
-    .patchNamespacedSecret(
-      appSecret?.body?.metadata?.name as string,
-      k8sNamespace,
-      secret,
-      undefined,
-      undefined,
-      undefined,
-      undefined,
-      { headers }
-    )
-    .then(
-      () => {
-        console.info('Successfully update the token');
-      },
-      (err) => {
-        console.warn(err, 'Failed to patch secret');
-      }
-    );
-}
-
-export = (app: Probot) => {
+export = (app: ProbotOnKube) => {
   // Respond to the GitHub app installation
   app.on('installation.created', async (context) => {
     // Get the installation token  and expiry time from the installation
@@ -113,29 +11,19 @@ export = (app: Probot) => {
     })) as InstallationAccessTokenAuthentication;
     const repos = context.payload.repositories;
     const orgName = repos[0].full_name.split('/')[0];
-    app.log.info(`Operating on the Github Org: ${orgName}`);
 
     // Iterate over the list of repositories for .github repo
     const repo_exist = Boolean(repos.find((r) => r.name === '.github'));
 
-    // TODO: Do exception handling for not having perm for repo creation.
     if (!repo_exist) {
-      app.log.info(
-        'Peribolos as services requires .github repo. \
-                    \nDidnt find .github repo. \
-                    \nCreating  the repository for app workflow.'
-      );
-      context.octokit.repos.createInOrg({ org: orgName, name: '.github' }).then(
-        () => {
-          app.log.info('Created .github repository');
-        },
-        (err) => {
+      app.log.info("Creating '.github' repository.");
+      context.octokit.repos
+        .createInOrg({ org: orgName, name: '.github' })
+        .catch((err) => {
           app.log.warn(err, 'Error creating repository');
-        }
-      );
+        });
     }
 
-    // TODO: check if secret exists with try catch
     const secret: V1Secret = {
       metadata: {
         name: 'peribolos-' + context.payload.installation.id,
@@ -151,7 +39,7 @@ export = (app: Probot) => {
         orgName: orgName,
       },
     };
-    await k8sApi.createNamespacedSecret(k8snamespace, secret);
+    await app.kube.core.createNamespacedSecret(app.kube.namespace, secret);ace, secret);
 
     // Create task run
     var payload = {
@@ -175,10 +63,10 @@ export = (app: Probot) => {
 
     // No need to check for token expiration since it was just created
     try {
-      await k8sCustomResourceApi.createNamespacedCustomObject(
+      await app.kube.custom.createNamespacedCustomObject(
         'tekton.dev',
         'v1beta1',
-        k8snamespace,
+        app.kube.namespace,
         'taskruns',
         payload
       );
@@ -226,10 +114,10 @@ export = (app: Probot) => {
       },
     };
     try {
-      await k8sCustomResourceApi.createNamespacedCustomObject(
+      await app.kube.custom.createNamespacedCustomObject(
         'tekton.dev',
         'v1beta1',
-        k8sNamespace,
+        app.kube.namespace,
         'taskruns',
         payload
       );
@@ -241,6 +129,6 @@ export = (app: Probot) => {
   // Respond to the GitHub app installation deleted
   app.on('installation.deleted', async (context) => {
     const name = 'peribolos-' + context.payload.installation.id;
-    await k8sApi.deleteNamespacedSecret(name, k8sNamespace);
+    await app.kube.core.deleteNamespacedSecret(name, app.kube.namespace);
   });
 };
